Rename components and init a new form component in order to put the inputfield and button there
