+++ src/components/simple/Core/TopBar/Form/Form.jsx
@@ -1,30 +1,25 @@
 import { useState } from "react";
+import InputField from "./InputField";
+import HistoryButton from "../../../../base/HistoryButton";
+import SearchButton from "../../../../base/SearchButton";
+import handleGetData from "../../../../../services/getData";
 
-export default function Form({ selectedCity, onSelectCity, onSubmitForm }) {
-  function enterValue(value) {
-    onSelectCity(value);
-  }
-
-  function submitForm(pressedKey) {
-    if (pressedKey === "Enter") {
-      onSubmitForm(selectedCity);
-    }
-  }
+export default function Form({ onClickHistory }) {
+  const [selectedCity, setSelectedCity] = useState("");
 
   return (
-    <>
-      <input
-        type="text"
-        placeholder="Search Location..."
-        spellCheck="false"
-        id="search-text"
-        maxLength="16"
-        value={selectedCity}
-        onChange={(e) => enterValue(e.target.value)}
-        onKeyUp={(e) => {
-          submitForm(e.key);
+    <form className="search-bar-container blank-line">
+      <InputField
+        selectedCity={selectedCity}
+        onSelectCity={setSelectedCity}
+        onSubmitForm={handleGetData}
+      />
+      <HistoryButton
+        onSelectButton={() => {
+          onClickHistory();
         }}
       />
-    </>
+      <SearchButton onSelectButton={handleGetData} city={selectedCity} />
+    </form>
   );
 }
