+++ src/components/simple/Core/TopBar/Form/InputField/index.js
@@ -1,0 +1,4 @@
+import InputField from './InputField.jsx';
+import './style.scss';
+
+export default InputField;
