+++ src/components/simple/Core/TopBar/Form/InputField/style.scss
@@ -1,0 +1,51 @@
+
+@media only screen and (max-width: 639px) { 
+  #search-text {
+    display: flex;
+    font-family: "Noto Sans", sans-serif;
+    font-optical-sizing: auto;
+    font-weight: 600;
+    font-size: 9px;
+    opacity: 0.8;
+    margin-bottom: 4x;
+    border-radius: 40px;
+    border-width: 3px;
+    width: 20vw;
+    padding: 3px;
+
+  }
+}
+
+
+@media only screen and (min-width: 640px)
+and (max-width: 1149px) {
+  #search-text {
+    display: flex;
+    flex-shrink: 1;
+    font-family: "Noto Sans", sans-serif;
+    font-optical-sizing: auto;
+    font-weight: 500;
+    font-size: 20px;
+    opacity: 0.8;
+    border-radius: 40px;
+    border-width: 3px;
+    width: 22vw;
+    padding-left: 5px;
+}
+} 
+
+
+@media only screen and (min-width: 1150px) {
+  #search-text {
+    display: flex;
+    font-family: "Noto Sans", sans-serif;
+    font-optical-sizing: auto;
+    font-weight: 500;
+    font-size: 18px;
+    opacity: 0.8;
+    width: 16vw;
+    border-radius: 40px;
+    border-width: 4px;
+    padding: 8px;
+  }
+}
