+++ src/components/simple/Core/TopBar/Form/index.js
@@ -1,4 +1,4 @@
-import Form from './Form.jsx';
+import Form from './Form';
 import './style.scss';
 
 export default Form;
