--- src/components/simple/LeftSide/Aside/index.js
@@ -1,4 +1,0 @@
-import Aside from "./Aside";
-import "./style.scss";
-
-export default Aside;
