--- src/components/simple/LeftSide/Background/index.js
@@ -1,4 +1,0 @@
-import Background from "./Background";
-import "./style.scss";
-
-export default Background;
