+++ src/components/simple/LeftSide/BackgroundLayout/BackgroundLayout.jsx
@@ -1,4 +1,4 @@
-export default function Background({ onClickBackground }) {
+export default function BackgroundLayout({ onClickBackground }) {
   return (
     <div
       className="sidebar-background-container active"
