+++ src/components/simple/LeftSide/BackgroundLayout/index.js
@@ -1,0 +1,4 @@
+import BackgroundLayout from "./BackgroundLayout";
+import "./style.scss";
+
+export default BackgroundLayout;
