+++ src/components/simple/LeftSide/BackgroundLayout/style.scss
