+++ src/components/simple/LeftSide/HistoryTab/index.js
@@ -1,0 +1,4 @@
+import HistoryTab from "./HistoryTab";
+import "./style.scss";
+
+export default HistoryTab;
