+++ src/components/simple/LeftSide/HistoryTab/style.scss
