+++ src/components/simple/RightSide/Weather/style.scss
@@ -1,6 +1,3 @@
-
-
-
 @media only screen and (max-width: 639px) { 
   .weather-container {
     display: flex;
@@ -8,7 +5,7 @@
     align-items: end;
     height: 18vh;
     max-width: 80vw;
-}
+  }
 
   .weather-forecast-container {
       display: flex;
